# app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/suggestions/config": {
      "get": {
        "operationId": "ConfigController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Настройки"
          }
        },
        "summary": "Получение настроек (Categories, Notifications)",
        "tags": [
          "Config"
        ]
      },
      "put": {
        "operationId": "ConfigController_updateConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённые настройки"
          }
        },
        "summary": "Обновление настроек",
        "tags": [
          "Config"
        ]
      }
    },
...
    "/api/v1/suggestions/settings/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Шаблоны уведомлений"
          }
        },
        "summary": "Получение шаблонов уведомлений",
        "tags": [
          "Notifications"
        ]
      },
      "put": {
        "operationId": "NotificationsController_updateNotifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationTemplatesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённые шаблоны"
          }
        },
        "summary": "Обновление шаблонов уведомлений",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/v1/suggestions/{suggestionId}/comments": {
      "get": {
        "operationId": "CommentsController_findBySuggestionId",
        "parameters": [
          {
            "name": "suggestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список комментариев"
          }
        },
        "summary": "Team Chat — список комментариев",
        "tags": [
          "Comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "suggestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Созданный комментарий"
          },
          "400": {
            "description": "X-User-Id required"
          }
        },
        "summary": "Team Chat — отправка комментария",
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/suggestions/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsByIdController_update",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённый комментарий"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Редактирование комментария",
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsByIdController_delete",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Удаление комментария",
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/suggestions/bans": {
      "get": {
        "operationId": "BansController_getBannedUsers[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список забаненных"
          }
        },
        "summary": "Список забаненных пользователей",
        "tags": [
          "Bans"
        ]
      },
      "post": {
        "operationId": "BansController_banUser[0]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь забанен"
          }
        },
        "summary": "Ban пользователя",
        "tags": [
          "Bans"
        ]
      }
    },
    "/api/v1/suggestions/banned-users": {
      "get": {
        "operationId": "BansController_getBannedUsers[1]",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список забаненных"
          }
        },
        "summary": "Список забаненных пользователей",
        "tags": [
          "Bans"
        ]
      },
      "post": {
        "operationId": "BansController_banUser[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь забанен"
          }
        },
        "summary": "Ban пользователя",
        "tags": [
          "Bans"
        ]
      }
    },
    "/api/v1/suggestions/bans/{userId}": {
      "delete": {
        "operationId": "BansController_unbanUser[0]",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь разбанен"
          }
        },
        "summary": "Unban пользователя",
        "tags": [
          "Bans"
        ]
      }
    },
    "/api/v1/suggestions/banned-users/{userId}": {
      "delete": {
        "operationId": "BansController_unbanUser[1]",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь разбанен"
          }
        },
        "summary": "Unban пользователя",
        "tags": [
          "Bans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Suggestions API",
    "description": "API для управления предложениями (suggestions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UpdateConfigDto": {
        "type": "object",
        "properties": {}
      },
      "NotificationTemplatesDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "BanUserDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}